added more structure to particles_between.cpp
Solved the problem of how to handle the double loop and built the code structure. Introduced the syntax for adding particles to the trajectory list. All that is needed now is constructors and other standard required methods, and the actual math algorithm for determining if a particle is between two other particles.